Add comm_plan traitlet to VegaFusionWidget
