+++ vegafusion-core/src/planning/mod.rs
@@ -1,3 +1,4 @@
 pub mod data_graph;
 pub mod extract;
 pub mod stitch;
+pub mod watch;
